Playback is now scheduled from the render thread
The rationale for doing this is that the AUFilePlayer has no callbacks
to tell when it loops or stops. I would like to add notifications when
these events happen. In order to do this I need to know the exact
timestamp at the start of playback. From here I can calculate the
timestamp when the file will stop playback (or loop) and schedule
events accordingly.